requirements.txt
pillow==10.4.0
torch==2.4.1
torchvision==0.19.1
numpy==1.24.4